% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr.R
\name{add_query_filter}
\alias{add_query_filter}
\title{add_query_filter}
\usage{
add_query_filter(q, filter)
}
\arguments{
\item{q}{A query object.}

\item{filter}{A filter object.}
}
\value{
A query body
}
\description{
This function binds a filter to a query body.
}
