% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr.R
\name{add_source}
\alias{add_source}
\title{add_source}
\usage{
add_source(q, source)
}
\arguments{
\item{q}{query body}

\item{source}{a vector contain source field names}
}
\value{
A query body
}
\description{
This function adds a source field to a query body in R obj
}
