% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr.R
\name{countRegionQuery}
\alias{countRegionQuery}
\title{Count SNPs based on specified chromosome, start position, end position and filter arguments.}
\usage{
countRegionQuery(
  chromosome_identifier,
  start_position = 1,
  end_position = 1e+05,
  filter_fields = NULL,
  search_hrc = FALSE
)
}
\arguments{
\item{chromosome_identifier}{The chromosome number (or 'X' for the X-chromosome)}

\item{start_position}{Start position region of search (default: 1)}

\item{end_position}{End position region of search (default: 100000)}

\item{filter_fields}{SNP attribute labels that should not be empty for the record to be retrieved}

\item{search_hrc}{Restrict results to variants mapped to the HRC r1.1 panel,
in hg19 coordinates. Defaults to \code{FALSE}.}
}
\value{
The count of SNPs matching the criteria.
}
\description{
Count SNPs based on specified chromosome, start position, end position and filter arguments.
}
\examples{
# Count SNPs on chromosome 1 between positions 10000 and 20000
count <- countRegionQuery(
  chromosome_identifier = "1",
  start_position = 10000,
  end_position = 20000
)
print(paste("Number of SNPs:", count))

# Count with filter fields
count <- countRegionQuery(
  chromosome_identifier = "X",
  start_position = 100000,
  end_position = 200000,
  filter_fields = c("pos", "chr")
)
print(paste("Number of SNPs:", count))

}
