% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr_newAPI.R
\name{create_annotations_query_string}
\alias{create_annotations_query_string}
\title{create_annotations_query_string}
\usage{
create_annotations_query_string(annotations)
}
\arguments{
\item{annotations}{A character vector of fields.}
}
\value{
A string formatted for a GraphQL query.
}
\description{
Convert a list of fields into a GraphQL query format.
}
