% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr.R
\name{exists_filter}
\alias{exists_filter}
\title{exists_filter}
\usage{
exists_filter(key)
}
\arguments{
\item{key}{A string, name for that field.}
}
\value{
A filter
}
\description{
This function returns filter depends on the existance of a certain field
}
