% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr.R
\name{keywordsQuery}
\alias{keywordsQuery}
\title{keywordsQuery}
\usage{
keywordsQuery(keywords)
}
\arguments{
\item{keywords}{keywords a string}

\item{configFile}{optional parameter that provide fields to return, FALSE by default}
}
\value{
A list contain variants
}
\description{
Perform full text search in our variant annotation dataset
}
