% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr_newAPI.R
\name{perform_graphql_query}
\alias{perform_graphql_query}
\title{perform_graphql_query}
\usage{
perform_graphql_query(query)
}
\arguments{
\item{query}{A character string representing the GraphQL query.}
}
\value{
A character string with the query result.
}
\description{
Perform a GraphQL query.
}
