% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr.R
\name{perform_search_find_count}
\alias{perform_search_find_count}
\title{perform_search}
\usage{
perform_search(q)
}
\arguments{
\item{q}{A qeury string}
}
\value{
A list contain variants
}
\description{
execute http request with a query json string
}
