% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr.R
\name{query_obj_to_json}
\alias{query_obj_to_json}
\title{query_obj_to_json}
\usage{
query_obj_to_json(q)
}
\arguments{
\item{q}{A qeury obj}
}
\value{
A json string.
}
\description{
This function convert a query obj to a json string.
}
