% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr.R
\name{range_filter}
\alias{range_filter}
\title{range_filter}
\usage{
range_filter(key, gt, lt)
}
\arguments{
\item{key}{A string, name for that field.}

\item{gt}{A number, optional, open lower boundary for that field}

\item{lt}{A number, optional, open upper boundary for that field}
}
\value{
A filter
}
\description{
This function return a filter to choose a certain numeric field within a range.
}
