% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr.R
\name{read_config}
\alias{read_config}
\title{read_config}
\usage{
read_config(file_name)
}
\arguments{
\item{file_name}{A string, full path to that configure file.}
}
\value{
A vector contain source fiedl names
}
\description{
Reading source configure from a file
}
