% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr_newAPI.R
\name{regionQ}
\alias{regionQ}
\title{regionQuery}
\usage{
regionQ(chr, start, end, annotations_to_retrieve)
}
\arguments{
\item{chr}{A character string for the chromosome.}

\item{start}{An integer for the start position.}

\item{end}{An integer for the end position.}

\item{annotations_to_retrieve}{A character vector of fields to retrieve.}
}
\value{
A data frame with the SNPs.
}
\description{
Get SNPs by region.
}
