% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr.R
\name{regionQuery}
\alias{regionQuery}
\title{Search for SNPs by chromosome id and position range.}
\usage{
regionQuery(
  chromosome_identifier,
  start_position = 1,
  end_position = 1e+05,
  fields = NULL,
  filter_fields = NULL,
  pagination_from = 0,
  pagination_size = 1000,
  fetch_all = FALSE,
  search_hrc = FALSE
)
}
\arguments{
\item{chromosome_identifier}{Chromosome id to search (e.g., "1", "2", "X")}

\item{start_position}{Start position region of search (default: 1)}

\item{end_position}{End position region of search (default: 100000)}

\item{fields}{Fields to return, can be JSON string, file path, or vector of attributes. Number of fields is limited to 20.}

\item{filter_fields}{SNP attribute labels that should not be empty for the record to be retrieved}

\item{pagination_from}{Pagination start index (default: 0)}

\item{pagination_size}{Pagination page size (default: 1000)}

\item{fetch_all}{If TRUE, retrieves all matching SNPs by downloading all pages (default: FALSE)}

\item{search_hrc}{Restrict results to variants mapped to the HRC r1.1 panel,
in hg19 coordinates. Defaults to \code{FALSE}.}
}
\value{
A list containing the SNP information.
}
\description{
Search for SNPs by chromosome id and position range.
}
\details{
If fetch_all is TRUE, pagination_from and pagination_size are ignored.
The function will return all matching SNPs in a single list.
It only supports up to 1,000,000 SNPs being fetched in total.

If using pagination (fetch_all=FALSE), you cannot fetch more than the first 10,000 SNPs over all pages.
pagination_from + pagination_size must be <= 10,000.
}
\examples{
# Search for SNPs on chromosome 1 between positions 10000 and 20000
snps <- regionQuery(
  chromosome_identifier = "1",
  start_position = 10000,
  end_position = 20000
)
print(snps)

# Search with specific fields returned
snps <- regionQuery(
  chromosome_identifier = "X",
  start_position = 100000,
  end_position = 200000,
  fields = c("Basic Info", "chr", "pos", "ref", "alt")
)
print(snps)

# Fetch all results
snps <- regionQuery(
  chromosome_identifier = "1",
  start_position = 10000,
  end_position = 20000,
  fetch_all = TRUE
)
print(snps)

}
