% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr_newAPI.R
\name{rsidQ}
\alias{rsidQ}
\title{rsidQuery}
\usage{
rsidQ(rsID, annotations_to_retrieve)
}
\arguments{
\item{rsID}{A character string for the rsID.}

\item{annotations_to_retrieve}{A character vector of fields to retrieve.}
}
\value{
A data frame with the SNPs.
}
\description{
Get SNP by rsID.
}
