% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr.R
\name{rsidQuery}
\alias{rsidQuery}
\title{rsidQuery}
\usage{
rsidQuery(rsid)
}
\arguments{
\item{rsid}{rsid a string}

\item{configFile}{optional parameter that provide fields to return, FALSE by default}
}
\value{
One variant if found
}
\description{
This function use a rsid to query a variant
}
