% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr_newAPI.R
\name{rsidsQ}
\alias{rsidsQ}
\title{rsidsQuery}
\usage{
rsidsQ(rsIDs, annotations_to_retrieve)
}
\arguments{
\item{rsIDs}{A character vector of rsIDs.}

\item{annotations_to_retrieve}{A character vector of fields to retrieve.}
}
\value{
A data frame with the SNPs.
}
\description{
Get SNPs by multiple rsIDs.
}
