% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr.R
\name{snpAttributesQuery}
\alias{snpAttributesQuery}
\title{Retrieve available list of SNP attributes.}
\usage{
snpAttributesQuery()
}
\value{
A list containing the available SNP attributes.
}
\description{
Retrieve available list of SNP attributes.
}
\examples{
# Retrieve all available SNP attributes
attributes <- snpAttributesQuery()
print(attributes)

}
