% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr.R
\name{term_filter}
\alias{term_filter}
\title{term_filter}
\usage{
term_filter(key, value)
}
\arguments{
\item{key}{A string, name for that field.}

\item{value}{A string, value for that field.}
}
\value{
A filter
}
\description{
This function returns a filter to choose a certain field is exactly some value
}
