% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr.R
\name{geneQuery}
\alias{geneQuery}
\title{Search for specified gene product; this can be a gene id, gene symbol or UniProt id.}
\usage{
geneQuery(gene = NULL, fields = NULL, filter_fields = NULL)
}
\arguments{
\item{gene}{Gene product to search}

\item{fields}{Fields to return, can be JSON string, file path, or vector of attributes}

\item{filter_fields}{SNP attribute labels that should not be empty for the record to be retrieved}
}
\value{
A list containing the SNP information.
}
\description{
Search for specified gene product; this can be a gene id, gene symbol or UniProt id.
}
\examples{
# Search for SNPs associated with a specific gene
gene_results <- geneQuery(gene = "BRCA1")
print(gene_results)

# Search with specific fields returned
gene_results <- geneQuery(gene = "TP53",
                         fields = c("Basic Info", "chr", "pos", "ref", "alt"))
print(gene_results)

}
