% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionQuery.R
\name{regionQuery}
\alias{regionQuery}
\title{perform Region query}
\usage{
regionQuery(contig, start, end)
}
\arguments{
\item{contig}{contig string}

\item{start}{sgtart position in int}

\item{end}{end position in int}
}
\value{
A dataframe contain variants
}
\description{
This function use a genome coordinate to query variants within that region
}
