% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr.R
\name{regionQuery}
\alias{regionQuery}
\title{Search for SNPs by chromosome id and position range.}
\usage{
regionQuery(
  chromosome_identifier,
  start_position = NULL,
  end_position = NULL,
  fields = NULL,
  filter_fields = NULL
)
}
\arguments{
\item{chromosome_identifier}{Chromosome id to search (e.g., "1", "2", "X")}

\item{start_position}{Start position region of search (default: 1)}

\item{end_position}{End position region of search (default: 100000)}

\item{fields}{Fields to return, can be JSON string, file path, or vector of attributes}

\item{filter_fields}{SNP attribute labels that should not be empty for the record to be retrieved}
}
\value{
A list containing the SNP information.
}
\description{
Search for SNPs by chromosome id and position range.
}
\examples{
# Search for SNPs on chromosome 1 between positions 10000 and 20000
snps <- regionQuery(chromosome_identifier = "1", 
                   start_position = 10000, 
                   end_position = 20000)
print(snps)

# Search with specific fields returned
snps <- regionQuery(chromosome_identifier = "X", 
                   start_position = 100000, 
                   end_position = 200000,
                   fields = c("Basic Info", "chr", "pos", "ref", "alt"))
print(snps)

}
