% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr.R
\name{regionQuery}
\alias{regionQuery}
\title{regionQuery}
\usage{
regionQuery(contig, start, end)
}
\arguments{
\item{contig}{contig string}

\item{start}{sgtart position in int}

\item{end}{end position in int}

\item{configFile}{optional parameter that provide fields to return, FALSE by default}
}
\value{
A list contain variants
}
\description{
This function use a genome coordinate to query variants within that region
}
