% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoqr.R
\name{rsidsQuery}
\alias{rsidsQuery}
\title{Search for specified list of RSIDs.}
\usage{
rsidsQuery(rsid_list = NULL, fields = NULL, filter_fields = NULL)
}
\arguments{
\item{rsid_list}{List of RSIDs to search, can be comma-separated string or vector of strings}

\item{fields}{Fields to return, can be JSON string, file path, or vector of attributes}

\item{filter_fields}{SNP attribute labels that should not be empty for the record to be retrieved}
}
\value{
A list containing the SNP information.
}
\description{
Search for specified list of RSIDs.
}
\examples{
# Search for specific RSIDs
rsid_results <- rsidsQuery(rsid_list = c("rs123456", "rs789012"))
print(rsid_results)

# Search with specific fields returned
rsid_results <- rsidsQuery(rsid_list = "rs123456,rs789012",
                          fields = c("Basic Info", "chr", "pos"))
print(rsid_results)

}
